Separated header component into header and home component
