+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './components/header/home.component';
+import { HomeComponent } from './components/header/header.component';
 import { FooterComponent } from './components/footer/footer.component';
 
 const routes: Routes = [
