+++ src/app/components/header/header.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+
+@Component({
+  selector: 'app-header',
+  templateUrl: './header.component.html',
+  styleUrl: './header.component.css'
+})
+export class HomeComponent {
+
+}
