--- src/app/components/header/home.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-
-@Component({
-  selector: 'app-home',
-  templateUrl: './home.component.html',
-  styleUrl: './home.component.css'
-})
-export class HomeComponent {
-
-}
